+++ src/TitleDetails/PiecesList/PiecesList.js
@@ -1,11 +1,11 @@
 import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
+import { without } from 'lodash';
 
 import {
   Icon,
   NoValue,
-  Tooltip,
 } from '@folio/stripes/components';
 import {
   acqRowFormatter,
@@ -15,7 +15,6 @@ import {
 } from '@folio/stripes-acq-components';
 
 import { PIECE_COLUMNS } from '../constants';
-import styles from './PiecesList.css';
 
 const sorters = {
   [PIECE_COLUMNS.enumeration]: ({ enumeration }) => enumeration?.toLowerCase(),
@@ -31,6 +30,7 @@ const columnMapping = {
   format: <FormattedMessage id="ui-receiving.piece.format" />,
   [PIECE_COLUMNS.receiptDate]: <FormattedMessage id="ui-receiving.piece.receiptDate" />,
   [PIECE_COLUMNS.receivedDate]: <FormattedMessage id="ui-receiving.piece.receivedDate" />,
+  [PIECE_COLUMNS.comment]: <FormattedMessage id="ui-receiving.piece.comment" />,
   request: <FormattedMessage id="ui-receiving.piece.request" />,
   selection: null,
 };
@@ -39,31 +39,10 @@ const formatter = {
   [PIECE_COLUMNS.copyNumber]: piece => piece.copyNumber || <NoValue />,
   [PIECE_COLUMNS.caption]: piece => piece.caption || <NoValue />,
   [PIECE_COLUMNS.enumeration]: piece => piece.enumeration || <NoValue />,
-  format: piece => (
-    <span>
-      {PIECE_FORMAT_LABELS[piece.format]}
-      {
-        Boolean(piece.comment) && (
-          <Tooltip
-            id={`piece-comment-${piece.id}`}
-            text={piece.comment}
-          >
-            {({ ref, ariaIds }) => (
-              <Icon
-                ref={ref}
-                size="small"
-                icon="comment"
-                ariaLabel={ariaIds.text}
-                iconClassName={styles.pieceComment}
-              />
-            )}
-          </Tooltip>
-        )
-      }
-    </span>
-  ),
+  format: piece => PIECE_FORMAT_LABELS[piece.format],
   [PIECE_COLUMNS.receiptDate]: piece => <FolioFormattedDate value={piece.receiptDate} />,
   [PIECE_COLUMNS.receivedDate]: piece => <FolioFormattedDate value={piece.receivedDate} />,
+  [PIECE_COLUMNS.comment]: piece => piece.comment || <NoValue />,
   barcode: piece => piece.barcode || <NoValue />,
   request: piece => (
     piece.request
@@ -76,6 +55,11 @@ const formatter = {
 const PiecesList = ({ pieces, id, visibleColumns, selectPiece, sortedColumn }) => {
   const hasRowClick = Boolean(selectPiece);
   const rowProps = useMemo(() => ({ alignLastColToEnd: hasRowClick }), [hasRowClick]);
+  const tableColumns = useMemo(() => (
+    pieces.some(({ comment }) => comment)
+      ? visibleColumns
+      : without(visibleColumns, 'comment')
+  ), [pieces, visibleColumns]);
 
   if (!pieces) return null;
 
@@ -88,7 +72,7 @@ const PiecesList = ({ pieces, id, visibleColumns, selectPiece, sortedColumn }) =
       interactive={false}
       rowFormatter={acqRowFormatter}
       rowProps={rowProps}
-      visibleColumns={visibleColumns}
+      visibleColumns={tableColumns}
       onRowClick={selectPiece}
       sortedColumn={sortedColumn}
       sorters={sorters}