+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Allow user to save and create another piece. Refs UIREC-188.
 * Tech debt: reducing code smells, usage of consts and resources from stripes-acq. Refs UIREC-192.
 * Validate user has made a change to the piece form before activating "Save" or "Quick receive" buttons. Refs UIREC-190.
+* Display comment in receiving in column rather than on hover. Refs UIREC-193.
 
 ## [2.0.2](https://github.com/folio-org/ui-receiving/tree/v2.0.2) (2021-11-04)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v2.0.1...v2.0.2)